Refactor buildRun reconciler class
Move some functions into new or existing classes under
the internal resource package.
